Refactor CSS and HTML for buttons
The commit includes clearer organization and naming changes in the CSS file, by renaming --btn-p-4 to --4px under a new Utils section. It also modifies the styles of the .base, .outline, and .solid buttons for consistency and clarity. Changes extend to the buttons-01 HTML file, with some button classes rearranged.