Update color mode button styling
The color mode button container class was renamed from "dark-light-container" to "color-mode-btn-container" and new styles were added for it and its button. Also, the color and size of the button's icon were changed, and a "scrolled" class is added to the icon when the window is scrolled for color change.